+++ hassio/src/ansi-to-html.js
@@ -1,68 +1,66 @@
-import { html } from "@polymer/polymer/lib/utils/html-tag";
+import { css } from "lit-element";
 
-export const ANSI_HTML_STYLE = html`
-  <style>
-    .bold {
-      font-weight: bold;
-    }
-    .italic {
-      font-style: italic;
-    }
-    .underline {
-      text-decoration: underline;
-    }
-    .strikethrough {
-      text-decoration: line-through;
-    }
-    .underline.strikethrough {
-      text-decoration: underline line-through;
-    }
-    .fg-red {
-      color: rgb(222, 56, 43);
-    }
-    .fg-green {
-      color: rgb(57, 181, 74);
-    }
-    .fg-yellow {
-      color: rgb(255, 199, 6);
-    }
-    .fg-blue {
-      color: rgb(0, 111, 184);
-    }
-    .fg-magenta {
-      color: rgb(118, 38, 113);
-    }
-    .fg-cyan {
-      color: rgb(44, 181, 233);
-    }
-    .fg-white {
-      color: rgb(204, 204, 204);
-    }
-    .bg-black {
-      background-color: rgb(0, 0, 0);
-    }
-    .bg-red {
-      background-color: rgb(222, 56, 43);
-    }
-    .bg-green {
-      background-color: rgb(57, 181, 74);
-    }
-    .bg-yellow {
-      background-color: rgb(255, 199, 6);
-    }
-    .bg-blue {
-      background-color: rgb(0, 111, 184);
-    }
-    .bg-magenta {
-      background-color: rgb(118, 38, 113);
-    }
-    .bg-cyan {
-      background-color: rgb(44, 181, 233);
-    }
-    .bg-white {
-      background-color: rgb(204, 204, 204);
-    }
-  </style>
+export const ANSI_HTML_STYLE = css`
+  .bold {
+    font-weight: bold;
+  }
+  .italic {
+    font-style: italic;
+  }
+  .underline {
+    text-decoration: underline;
+  }
+  .strikethrough {
+    text-decoration: line-through;
+  }
+  .underline.strikethrough {
+    text-decoration: underline line-through;
+  }
+  .fg-red {
+    color: rgb(222, 56, 43);
+  }
+  .fg-green {
+    color: rgb(57, 181, 74);
+  }
+  .fg-yellow {
+    color: rgb(255, 199, 6);
+  }
+  .fg-blue {
+    color: rgb(0, 111, 184);
+  }
+  .fg-magenta {
+    color: rgb(118, 38, 113);
+  }
+  .fg-cyan {
+    color: rgb(44, 181, 233);
+  }
+  .fg-white {
+    color: rgb(204, 204, 204);
+  }
+  .bg-black {
+    background-color: rgb(0, 0, 0);
+  }
+  .bg-red {
+    background-color: rgb(222, 56, 43);
+  }
+  .bg-green {
+    background-color: rgb(57, 181, 74);
+  }
+  .bg-yellow {
+    background-color: rgb(255, 199, 6);
+  }
+  .bg-blue {
+    background-color: rgb(0, 111, 184);
+  }
+  .bg-magenta {
+    background-color: rgb(118, 38, 113);
+  }
+  .bg-cyan {
+    background-color: rgb(44, 181, 233);
+  }
+  .bg-white {
+    background-color: rgb(204, 204, 204);
+  }
 `;
 
 export function parseTextToColoredPre(text) {
